requirements.txt
pytest==7.4.2
pytest-cov==4.1.0
pylint==3.0.1
black==23.10.0
fire==0.5.0
fastapi==0.104.0
uvicorn==0.23.2
pika==1.3.2
python-decouple==3.8
httpx==0.25.0
CouchDB==1.2